requirements.txt
Flask
flask-cors
pymongo
python-dotenv
google
google-generativeai
python-dateutil
Werkzeug
opencv-python-headless
pytesseract
pdf2image
pyzbar
Pillow